Restructured Plugin Loader
Restructured the way in which plugins are loaded. Plugins are now passed the form and config objects and draw directly to the form. They are also responsible for saving their own elements. A compile-time depency bug was also fixed. The code in this commit compiles and runs, but is non-functional.